+++ packages/sync-providers/src/file-based/onedrive/onedrive.model.ts
@@ -1,0 +1,28 @@
+export interface OneDrivePrivateCfg {
+  encryptKey?: string;
+  useCustomApp?: boolean;
+  clientId: string;
+  tenantId: string;
+  syncFolderPath?: string;
+  accessToken?: string;
+  refreshToken?: string;
+  tokenExpiresAt?: number;
+}
+
+export interface OneDriveItem {
+  id: string;
+  name: string;
+  eTag?: string;
+  folder?: Record<string, unknown>;
+  file?: Record<string, unknown>;
+}
+
+export interface OneDriveListResponse {
+  value?: OneDriveItem[];
+}
+
+export interface OneDriveTokenResponse {
+  access_token: string;
+  refresh_token?: string;
+  expires_in: number;
+}
