+++ packages/sync-providers/src/onedrive.ts
@@ -1,0 +1,11 @@
+export {
+  OneDrive,
+  PROVIDER_ID_ONEDRIVE,
+  type OneDriveDeps,
+} from './file-based/onedrive/onedrive';
+export type {
+  OneDrivePrivateCfg,
+  OneDriveItem,
+  OneDriveListResponse,
+  OneDriveTokenResponse,
+} from './file-based/onedrive/onedrive.model';
