+++ packages/sync-providers/tsup.config.ts
@@ -6,6 +6,7 @@ export default defineConfig({
     'src/webdav.ts',
     'src/local-file.ts',
     'src/super-sync.ts',
+    'src/onedrive.ts',
     'src/http.ts',
     'src/errors.ts',
     'src/credential-store.ts',
