+++ src/app/op-log/sync-providers/file-based/onedrive/onedrive.model.ts
@@ -1,29 +1,6 @@
-import { SyncProviderPrivateCfgBase } from '../../../core/types/sync.types';
-
-export interface OneDrivePrivateCfg extends SyncProviderPrivateCfgBase {
-  useCustomApp?: boolean;
-  clientId: string;
-  tenantId: string;
-  syncFolderPath?: string;
-  accessToken?: string;
-  refreshToken?: string;
-  tokenExpiresAt?: number;
-}
-
-export interface OneDriveItem {
-  id: string;
-  name: string;
-  eTag?: string;
-  folder?: Record<string, unknown>;
-  file?: Record<string, unknown>;
-}
-
-export interface OneDriveListResponse {
-  value?: OneDriveItem[];
-}
-
-export interface OneDriveTokenResponse {
-  access_token: string;
-  refresh_token?: string;
-  expires_in: number;
-}
+export type {
+  OneDrivePrivateCfg,
+  OneDriveItem,
+  OneDriveListResponse,
+  OneDriveTokenResponse,
+} from '@sp/sync-providers/onedrive';
